requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.4
mysqlclient==2.2.6
pillow==11.1.0
python-dotenv==1.0.1
reportlab==4.2.5
sqlparse==0.5.3
tzdata==2024.2
